+++ testkit/examples/timestamping/main.rs
@@ -18,6 +18,9 @@ extern crate exonum_testkit;
 extern crate serde_derive;
 #[macro_use]
 extern crate exonum_derive;
+
+use std::borrow::Cow;
+
 use exonum::{
     api::node::public::explorer::{BlocksQuery, BlocksRange, TransactionQuery},
     blockchain::{